feat: revitalize benches for smartcore 0.6.x + wire CI perf gate
Stage 0 + 1: bump deps to match smartcore 0.6.x (edition 2024, MSRV 1.85,
ndarray 0.17, rand 0.10, criterion 0.5, itertools 0.15); fix all legacy
benches (distance, fastpair, linear, naive_bayes, svc) to compile and pass
fmt/clippy against the current API.

Stage 2: add dedicated benchmarks for all 5 hot paths flagged in smartcore#407:
  - matmul.rs       → Array2::matmul (arrays.rs:1117), {64,256,1024}²
  - ab.rs           → Array2::ab, 4 transpose combos (arrays.rs:1143)
  - svd.rs          → SVDDecomposable::svd (svd.rs:71), DenseMatrix + ndarray
  - cover_tree.rs   → CoverTree::new/find (cover_tree.rs:37) vs LinearKNNSearch
  - iterator_mut.rs → DenseMatrix::iterator_mut (matrix.rs:545) + MutView
                      (matrix.rs:254) — the #368 no-regression surface

Stage 3: add .github/workflows/bench.yml (criterion job) triggered by
repository_dispatch from smartcore and workflow_dispatch. smartcore is
resolved from crates.io (latest 0.6.x). Results persisted to gh-pages
via benchmark-action/github-action-benchmark.

Stage 5: add iai-callgrind deterministic instruction-count gate
(linux-only, Valgrind) mirroring the 5 hot-path benches, with a 120%
alert threshold and fail-on-alert. Posts a pass/fail status check back to
the triggering smartcore commit.

Stage 6: rewrite README (fix dead CONTRIBUTING link, document dispatch
contract, local-run, criterion-vs-iai rationale, required secrets);
add .github/workflows/lint.yml for harness validation on PRs.

Closes smartcorelib/smartcore#407 (smartcore-benches side).

+++ benches/naive_bayes.rs
@@ -1,5 +1,5 @@
 use criterion::BenchmarkId;
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use criterion::{Criterion, black_box, criterion_group, criterion_main};
 
 use ndarray::Array2;
 use smartcore::linalg::basic::arrays::Array2 as BaseArray2;
@@ -13,7 +13,7 @@ pub fn gaussian_naive_bayes_fit_benchmark(c: &mut Criterion) {
         for n_features in [10_usize, 100_usize, 1000_usize].iter() {
             let x = DenseMatrix::<f64>::rand(*n_samples, *n_features);
             let y: Vec<usize> = (0..*n_samples)
-                .map(|i| (i % *n_samples / 5_usize) as usize)
+                .map(|i| i % *n_samples / 5_usize)
                 .collect::<Vec<usize>>();
             group.bench_with_input(
                 BenchmarkId::from_parameter(format!(
