+++ benches/svc.rs
@@ -1,9 +1,9 @@
 use criterion::BenchmarkId;
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use criterion::{Criterion, black_box, criterion_group, criterion_main};
 use smartcore::linalg::basic::arrays::Array2 as BaseArray2;
 use smartcore::linalg::basic::matrix::DenseMatrix;
-use smartcore::svm::svc::{MultiClassSVC, SVCParameters};
 use smartcore::svm::Kernels;
+use smartcore::svm::svc::{MultiClassSVC, SVCParameters};
 
 pub fn multiclass_svc_fit_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("MultiClassSVC::fit");
@@ -12,7 +12,7 @@ pub fn multiclass_svc_fit_benchmark(c: &mut Criterion) {
         for n_features in [10_usize, 100_usize, 1000_usize].iter() {
             let x = DenseMatrix::<f64>::rand(*n_samples, *n_features);
             let y: Vec<usize> = (0..*n_samples)
-                .map(|i| (i % *n_samples / 5_usize) as usize)
+                .map(|i| i % *n_samples / 5_usize)
                 .collect::<Vec<usize>>();
             let parameters = SVCParameters::default()
                 .with_c(1.0)
